+++ sdk/nodejs/identityplatform/defaultSupportedIdpConfig.ts
@@ -47,6 +47,11 @@ export class DefaultSupportedIdpConfig extends pulumi.CustomResource {
      */
     public readonly enabled!: pulumi.Output<boolean | undefined>;
     /**
+     * ID of the IDP. Possible values include: * 'apple.com' * 'facebook.com' * 'gc.apple.com' * 'github.com' *
+     * 'google.com' * 'linkedin.com' * 'microsoft.com' * 'playgames.google.com' * 'twitter.com' * 'yahoo.com'
+     */
+    public readonly idpId!: pulumi.Output<string>;
+    /**
      * The name of the DefaultSupportedIdpConfig resource
      */
     public /*out*/ readonly name!: pulumi.Output<string>;
@@ -71,6 +76,7 @@ export class DefaultSupportedIdpConfig extends pulumi.CustomResource {
             inputs["clientId"] = state ? state.clientId : undefined;
             inputs["clientSecret"] = state ? state.clientSecret : undefined;
             inputs["enabled"] = state ? state.enabled : undefined;
+            inputs["idpId"] = state ? state.idpId : undefined;
             inputs["name"] = state ? state.name : undefined;
             inputs["project"] = state ? state.project : undefined;
         } else {
@@ -81,9 +87,13 @@ export class DefaultSupportedIdpConfig extends pulumi.CustomResource {
             if (!args || args.clientSecret === undefined) {
                 throw new Error("Missing required property 'clientSecret'");
             }
+            if (!args || args.idpId === undefined) {
+                throw new Error("Missing required property 'idpId'");
+            }
             inputs["clientId"] = args ? args.clientId : undefined;
             inputs["clientSecret"] = args ? args.clientSecret : undefined;
             inputs["enabled"] = args ? args.enabled : undefined;
+            inputs["idpId"] = args ? args.idpId : undefined;
             inputs["project"] = args ? args.project : undefined;
             inputs["name"] = undefined /*out*/;
         }
@@ -115,6 +125,11 @@ export interface DefaultSupportedIdpConfigState {
      */
     readonly enabled?: pulumi.Input<boolean>;
     /**
+     * ID of the IDP. Possible values include: * 'apple.com' * 'facebook.com' * 'gc.apple.com' * 'github.com' *
+     * 'google.com' * 'linkedin.com' * 'microsoft.com' * 'playgames.google.com' * 'twitter.com' * 'yahoo.com'
+     */
+    readonly idpId?: pulumi.Input<string>;
+    /**
      * The name of the DefaultSupportedIdpConfig resource
      */
     readonly name?: pulumi.Input<string>;
@@ -141,6 +156,11 @@ export interface DefaultSupportedIdpConfigArgs {
      * If this IDP allows the user to sign in
      */
     readonly enabled?: pulumi.Input<boolean>;
+    /**
+     * ID of the IDP. Possible values include: * 'apple.com' * 'facebook.com' * 'gc.apple.com' * 'github.com' *
+     * 'google.com' * 'linkedin.com' * 'microsoft.com' * 'playgames.google.com' * 'twitter.com' * 'yahoo.com'
+     */
+    readonly idpId: pulumi.Input<string>;
     /**
      * The ID of the project in which the resource belongs.
      * If it is not provided, the provider project is used.