+++ sdk/nodejs/storage/bucketACL.ts
@@ -7,11 +7,6 @@ import * as outputs from "../types/output";
 import * as utilities from "../utilities";
 
 /**
- * Creates a new bucket ACL in Google cloud storage service (GCS). For more information see 
- * [the official documentation](https://cloud.google.com/storage/docs/access-control/lists) 
- * and 
- * [API](https://cloud.google.com/storage/docs/json_api/v1/bucketAccessControls).
- *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/r/storage_bucket_acl.html.markdown.
  */
 export class BucketACL extends pulumi.CustomResource {