+++ sdk/python/pulumi_gcp/iap/tunnel_instance_iam_member.py
@@ -12,52 +12,23 @@ from .. import utilities, tables
 class TunnelInstanceIAMMember(pulumi.CustomResource):
     condition: pulumi.Output[dict]
     etag: pulumi.Output[str]
-    """
-    (Computed) The etag of the instance's IAM policy.
-    """
     instance: pulumi.Output[str]
-    """
-    The name of the instance.
-    """
     member: pulumi.Output[str]
     project: pulumi.Output[str]
-    """
-    The ID of the project in which the resource belongs. If it
-    is not provided, the provider project is used.
-    """
     role: pulumi.Output[str]
-    """
-    The role that should be applied. Only one
-    `iap.TunnelInstanceIAMBinding` can be used per role. Note that custom roles must be of the format
-    `[projects|organizations]/{parent-name}/roles/{role-name}`.
-    """
     zone: pulumi.Output[str]
-    """
-    The zone of the instance. If
-    unspecified, this defaults to the zone configured in the provider.
-    """
     def __init__(__self__, resource_name, opts=None, condition=None, instance=None, member=None, project=None, role=None, zone=None, __props__=None, __name__=None, __opts__=None):
         """
         Create a TunnelInstanceIAMMember resource with the given unique name, props, and options.
         
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[str] instance: The name of the instance.
-        :param pulumi.Input[str] project: The ID of the project in which the resource belongs. If it
-               is not provided, the provider project is used.
-        :param pulumi.Input[str] role: The role that should be applied. Only one
-               `iap.TunnelInstanceIAMBinding` can be used per role. Note that custom roles must be of the format
-               `[projects|organizations]/{parent-name}/roles/{role-name}`.
-        :param pulumi.Input[str] zone: The zone of the instance. If
-               unspecified, this defaults to the zone configured in the provider.
         
         The **condition** object supports the following:
         
           * `description` (`pulumi.Input[str]`)
           * `expression` (`pulumi.Input[str]`)
           * `title` (`pulumi.Input[str]`)
-
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/r/iap_tunnel_instance_iam_member.html.markdown.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -104,23 +75,12 @@ class TunnelInstanceIAMMember(pulumi.CustomResource):
         :param str resource_name: The unique name of the resulting resource.
         :param str id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[str] etag: (Computed) The etag of the instance's IAM policy.
-        :param pulumi.Input[str] instance: The name of the instance.
-        :param pulumi.Input[str] project: The ID of the project in which the resource belongs. If it
-               is not provided, the provider project is used.
-        :param pulumi.Input[str] role: The role that should be applied. Only one
-               `iap.TunnelInstanceIAMBinding` can be used per role. Note that custom roles must be of the format
-               `[projects|organizations]/{parent-name}/roles/{role-name}`.
-        :param pulumi.Input[str] zone: The zone of the instance. If
-               unspecified, this defaults to the zone configured in the provider.
         
         The **condition** object supports the following:
         
           * `description` (`pulumi.Input[str]`)
           * `expression` (`pulumi.Input[str]`)
           * `title` (`pulumi.Input[str]`)
-
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/r/iap_tunnel_instance_iam_member.html.markdown.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))