+++ sdk/python/pulumi_gcp/monitoring/__init__.py
@@ -7,3 +7,4 @@ from .alert_policy import *
 from .group import *
 from .notification_channel import *
 from .uptime_check_config import *
+from .get_notification_channel import *